requirements.txt
PyYAML
requests-html
apscheduler
zmail
